Adds STENCILA_INLINE compilation directive
Some parts of Stencila don't require many dependencies (e.g. Frame, Query).
For these, it can be useful not to have to link in `libstencila.a` and to
just inline functions. By defining STENCILA_INLINE, method definitions are included
into headers. It then becomes the responsibility of the programmer to link to
third party libraries (e.g. Boost) as required.