Restore dependencies from spread data.
This allows a sheet to be viewed witout having to do an `update()` first
to calculate the dependency graph. The current approach is to store dependencies in the spread
variable `depends_`. A simpler alternative would be to redetermine dependencies each time
the sheet is opened but that would be more computationally expensive.