Embedded server address handling for (a) redirects (b) websocket connections
Using 'localhost' all the time is not great. In particular, when running
sessions in a local Docker container, a user may want to use the IP of the container, not
localhost (apparently, this i recommended by Docker). Two different approaches are required
for the cases (a) and (b) above.

For (a) a redirect using a relative URL is not possible because for
local component addresses (those that map directly to a filesystem path)
we use a leading forward slash and the double leading slashes in the redirect URL get
dropped by browsers. So, the embedded server needs to find out which
address it has been request on and redirect using that. Note that the
embedded server can be accepting connections on multiple IP addresses.

For (b), we need to get the client to construct the websocket URL because on the server side
that part of the code (in Component:call) currently doesn't know what IP the request
was made on (and probably never should).