feature: add parser keyword
This introduces the custom keyword `parser` to schema definitions.
Parsers allow users to enter shorter, and varying string formats for nodes
e.g. `John Tom` for an array of `givenNames` using the `ssv` parser.
To use these parsers schema authors need to add the `parser` keyword
and, where necessary write a parser function that
takes a spring and returns the desired node.
See `Person.ts` for an example.