fix(Article): Do not require authors and title
After using the `Article` type in Encoda extensively, it seems inappropriate to have these as required. The `Article` type is often used as a container for content. In Encoda, we often have to use an empty author list and an empty string, or 'Untitled' for title. Most of the time these requirements feel uncessary.