fix(*): Reorder property type alternatives for improved coercion
Previously we used an undocumented convention of ordering alternative types for a property (i.e `anyOf` rules) in order of increasing complexity (e.g. `string`, then `Date`). In Encoda we use Ajv to coerce data to the schema. Ajv attempts coercion in the order listed. This meant that if `string` was listed first, that no other coercion would be attempted.

This PR changes the order of types in various `anyOf` rules so that coercion is attempted in the "right" order.

There are associated changes to `Article.title` and `Article.description` to make coercion behave as expected.

There is an associated PR in Encoda.