fix(RPNG): Increase subtedly of affordance
This make the affordances on `CodeChunks` and `CodeExpressions` in RNGs much more subtle, showing only a green dot in the top left corner (middle left for expressions) and reducing whitespace around them. The intent is to create images that are close to those that will appear in a published article but with an indicator that something can be done with them.

In the furture we might use a red square instead for code elements that have an error on them.