feat(Custom errors): Define custom error types and publish them under the "errors" topic
By passing up details of the error we are able to
provide remidial action by the user. For the Node.js
bindings, the usual error message string is thrown, but
it is also possible to subscribe to the "errors" pubsub topic
to be able to provide notifications to  the user.