fix(TypeScript): Reduce number, and simplify, type guards
We where exporting quite a lot of type guard function, several of which
were effectively aliases  (e.g. `isListItem`) some of which effiectively duplicated
the functionality of others.

This commit attempts to simplify the guards and reduce them
to a mimimal, but still useful set. It also aims to reduce the confusion
between a guard for a specific type `isA` and a guard for a family of types
e.g. `isTypeOf`. The previous `isCreativeWork` was actually testing for waht is now
`isTypeOf('CreativeWork')(node)` but could be confused for `isA('CreativeWork', node)`.