fix(TypeScript bindings): Revert to using any
Whilst using `unknown` (or `Node`) would be better it breaks a lot of thing in Encoda.
I started to make some necessary fixes but it
was time consuming, particularly given the
other concurrent changes in progress.
With the view to getting some stability in the
shortterm, I am reverting this change, it
can be re-introduced later.