fix(CreativeWork): Allow an array of nodes
This was done because of the numerous breakages the narrower type had in Encoda.
Note that this property is actually already
narrowed in several places e.g. `Article.content`.
I suggest that we look at naroowing this later
perhaps even dropping this property in
favour of tighter typings on override properties.