feat(Desktop): Enable "New Project" workflow
Currently this function is almost identical to the `Open Project` command,
however it automatically opens a new untitled file in the editor panel once
the project window is ready.

In the future we'd like to explore a custom dialogue window allowing advanced
features such as selecting a project template, bootstrapping a project based
on GitHub repo or other file sources.