chore(Errors): Remove deprecated "sideband" error functions
The original intent here was to be able to associate non-fatal errors with user
actions in the desktop. But the use of the `attempt` function in Rust caused
confusion with  error propogation, particularly when using the CLI.
Furthermore, we alreading us`tracing:warn` and `tracing::error` for reporting
non-fatal errors so there was duplication and a lack of consistency on what
errors should go where.