fix(Documents): Remove option to patch node when executing it
Previously we gave the option to provide a patch when executing.
This was causing an issue because there was different logic around
setting the code language in the two places (using `ext` vs title case).
Separating executing and patching avoids this sort of issue and makes it easier to
reason about changes in state.