refactor(Execution plans): Rename "steps" to "tasks"
This change reduces the number of names we
use for things. We originally used `Step` instead of `Task` because we
already used `Task` in kernels. But this introduced a new name for
functionally the same thing.

In the future, we may find that it makes sense to merge the `Task`,
`TaskInfo` and `PlanTask` structs.