refactor(Images): Make  image layer determinism optional
A consequence of making layers deterministic (i.e. ignore modification time and other data)
was that generated images lost the modification times of files (and had odd times
e.g. 2006-06-29, not even 1970-01-01).  This change allows determinism to be optional
which in turn allows for images that retain all filesystem metadata.