chore(demo devcontainer): moved to separate repo
After some experimenting with improving startup times and UX
I have moved this demo devcontainer to https://github.com/stencila/demo-codespace.
This allows us to configure (and pay for) prebuilds of this container which dramatically
improve startup times. It is also likely to be easier to launch the Codespace for the demo
from Stencila Cloud via the GitHub REST API.

See https://github.com/stencila/stencila/pull/2373 for where this was introduced.