fix(web & vscode): load initial content into views for greater robustness
With this change, document previews in the browser and in VSCode,
begin with an initial version of HTML. This gives better immediate responsiveness,
for example when clicking on a View code lens to see the details of a node.
It should also give more robustness because the document will be available
before a websocket connection needs to be made.