refactor(codecs): use `csv` crate for lighter weight CSV/TSV support; make `codec-polars` optional
Polars is heavy and was increasing build times and binary sizes. Initially it was used because it supports
data types for columns out of the box. But we are able to achieve this typing relatively simply
with the much lighter weight `csv` crate. The `stencila-polars-crate` is retained for now but may
be removed in the future.